Move instances_from_po_lines function to digital_bookplates plugin
